LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/media-elements/interfaces/TextTrackCue/pauseOnExit-expected.txt

CONSOLE MESSAGE: Cross-origin text track load denied by Cross-Origin Resource Sharing policy.

PASS TextTrackCue.pauseOnExit, script-created cue 
FAIL TextTrackCue.pauseOnExit, parsed cue assert_unreached: got error event Reached unreachable code
